Fix specs
Not all specs were being run, because I accidentally commited a change
in spec/spec.js file. After enabling all of the specs it turned out that
2 of them are failing and it seems that it was caused by cloneNode on a
fragment. I'm not sure why was it added in the first place, but
cloneNode is needed on DocumentFragment only if a given fragment will be
reused which is not the case here